backend_version/requirements.txt
fastapi
uvicorn
numpy
opencv-python-headless
pillow
scikit-learn
matplotlib
scipy 